# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tuboshu Service Api",
        "title": "Tuboshu Service Api",
        "termsOfService": "https://github.com/opensourceai",
        "contact": {},
        "license": {},
        "version": "2.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
              "produces": [
                "application/json"
              ],
              "tags": [
                "Auth"
              ],
              "summary": "获取认证信息",
              "parameters": [
                {
                  "description": "user",
                  "name": "user",
                  "in": "body",
                  "required": true,
                  "schema": {
                    "$ref": "#/definitions/api.auth"
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "OK",
                  "schema": {
                    "$ref": "#/definitions/app.Response"
                  }
                },
                "500": {
                  "description": "Internal Server Error",
                  "schema": {
                    "$ref": "#/definitions/app.Response"
                  }
                }
              }
            }
        },
        "/auth/register": {
            "post": {
              "produces": [
                "application/json"
              ],
              "tags": [
                "Auth"
              ],
              "summary": "添加用户",
              "parameters": [
                {
                  "description": "user",
                  "name": "user",
                  "in": "body",
                  "required": true,
                  "schema": {
                    "$ref": "#/definitions/models.User"
                  }
                }
              ],
              "responses": {
                "200": {
                  "description": "OK",
                  "schema": {
                    "$ref": "#/definitions/app.Response"
                  }
                },
                "500": {
                  "description": "Internal Server Error",
                  "schema": {
                    "$ref": "#/definitions/app.Response"
                  }
                }
              }
            }
        },
      "/auth/test": {
        "get": {
          "security": [
            {
              "ApiKeyAuth": []
            }
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Auth"
          ],
          "summary": "auth测试",
          "parameters": [
            {
              "type": "string",
              "description": "string",
              "name": "str",
              "in": "query",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        }
      },
      "/v1/broad": {
        "get": {
          "produces": [
            "application/json"
          ],
          "tags": [
            "Board"
          ],
          "summary": "获取全部版块列表信息",
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        }
      },
      "/v1/broad/{id}": {
        "get": {
          "produces": [
            "application/json"
          ],
          "tags": [
            "Board"
          ],
          "summary": "获取某版块信息",
          "parameters": [
            {
              "type": "string",
              "description": "id",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        }
      },
      "/v1/broad/{id}/posts": {
        "get": {
          "produces": [
            "application/json"
          ],
          "tags": [
            "Board"
          ],
          "summary": "获取某版块的帖子",
          "parameters": [
            {
              "type": "string",
              "description": "id",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "type": "integer",
              "description": "页数",
              "name": "page",
              "in": "query"
            },
            {
              "type": "integer",
              "description": "条数",
              "name": "size",
              "in": "query"
            },
            {
              "type": "string",
              "description": "排序字段",
              "name": "order_by",
              "in": "query"
            },
            {
              "type": "string",
              "description": "升序, 降序",
              "name": "sorter",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        }
      },
      "/v1/comment": {
        "put": {
          "security": [
            {
              "ApiKeyAuth": []
            }
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Comment"
          ],
          "summary": "修改评论",
          "parameters": [
            {
              "description": "comment",
              "name": "comment",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/dto.CommentUpdate"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        },
        "post": {
          "security": [
            {
              "ApiKeyAuth": []
            }
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Comment"
          ],
          "summary": "新增评论",
          "parameters": [
            {
              "description": "comment",
              "name": "comment",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/models.Comment"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        },
        "delete": {
          "security": [
            {
              "ApiKeyAuth": []
            }
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Comment"
          ],
          "summary": "删除评论",
          "parameters": [
            {
              "description": "ids",
              "name": "ids",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/dto.Ids"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        }
      },
      "/v1/post": {
        "get": {
          "security": [
            {
              "ApiKeyAuth": []
            }
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Post"
          ],
          "summary": "获取用户自身帖子",
          "parameters": [
            {
              "type": "string",
              "description": "pageNum",
              "name": "page",
              "in": "query",
              "required": true
            },
            {
              "type": "string",
              "description": "pageSize",
              "name": "size",
              "in": "query",
              "required": true
            },
            {
              "type": "string",
              "description": "orderBy",
              "name": "orderBy",
              "in": "query"
            },
            {
              "type": "string",
              "description": "sorter",
              "name": "sorter",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        },
        "put": {
          "security": [
            {
              "ApiKeyAuth": []
            }
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Post"
          ],
          "summary": "修改用户自身帖子",
          "parameters": [
            {
              "description": "post",
              "name": "post",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/models.Post"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        },
        "post": {
          "security": [
            {
              "ApiKeyAuth": []
            }
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Post"
          ],
          "summary": "新增帖子",
          "parameters": [
            {
              "description": "post",
              "name": "post",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/models.Post"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        },
        "delete": {
          "security": [
            {
              "ApiKeyAuth": []
            }
          ],
          "produces": [
            "application/json"
          ],
          "tags": [
            "Post"
          ],
          "summary": "删除用户帖子",
          "parameters": [
            {
              "description": "postIds",
              "name": "postIds",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/v1.postIds"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        }
      },
      "/v1/post/{id}": {
        "get": {
          "produces": [
            "application/json"
          ],
          "tags": [
            "Post"
          ],
          "summary": "获取帖子信息",
          "parameters": [
            {
              "type": "string",
              "description": "id",
              "name": "id",
              "in": "path",
              "required": true
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        }
      },
      "/v1/post/{id}/comments": {
        "get": {
          "produces": [
            "application/json"
          ],
          "tags": [
            "Post"
          ],
          "summary": "获取帖子的评论",
          "parameters": [
            {
              "type": "string",
              "description": "ID",
              "name": "id",
              "in": "path",
              "required": true
            },
            {
              "type": "integer",
              "description": "页数",
              "name": "page",
              "in": "query"
            },
            {
              "type": "integer",
              "description": "条数",
              "name": "size",
              "in": "query"
            },
            {
              "type": "string",
              "description": "排序字段",
              "name": "order_by",
              "in": "query"
            },
            {
              "type": "string",
              "description": "升序, 降序",
              "name": "sorter",
              "in": "query"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            },
            "500": {
              "description": "Internal Server Error",
              "schema": {
                "$ref": "#/definitions/app.Response"
              }
            }
          }
        }
      }
    },
    "definitions": {
      "api.auth": {
        "type": "object",
        "properties": {
          "password": {
            "description": "密码",
            "type": "string"
          },
          "username": {
            "description": "用户名",
            "type": "string"
          }
        }
      },
      "app.Response": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "data": {
            "type": "object"
          },
          "msg": {
            "type": "string"
          }
        }
      },
      "dto.CommentUpdate": {
        "type": "object",
        "properties": {
          "comment_content": {
            "description": "评论的内容",
            "type": "string"
          },
          "id": {
            "description": "评论ID",
            "type": "integer"
          }
        }
      },
      "dto.Ids": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "帖子IDs",
            "type": "array",
            "items": {
              "type": "integer"
            }
          }
        }
      },
      "models.Comment": {
        "type": "object",
        "properties": {
          "comment_content": {
            "description": "评论内容",
            "type": "string"
          },
          "created_on": {
            "description": "新建时间",
            "type": "integer"
          },
          "deleted_on": {
            "description": "删除时间 当未被删除时,即为0,忽略此字段",
            "type": "integer"
          },
          "from_uid": {
            "description": "评论者ID",
            "type": "integer"
          },
          "id": {
            "description": "ID",
            "type": "integer"
          },
          "modified_on": {
            "description": "修改时间",
            "type": "integer"
          },
          "post_id": {
            "description": "主题ID",
            "type": "integer"
          },
          "to_uid": {
            "description": "被评论者ID",
            "type": "integer"
          }
        }
      },
      "models.Post": {
        "type": "object",
        "properties": {
          "board_id": {
            "description": "版块ID",
            "type": "integer"
          },
          "content": {
            "description": "内容",
            "type": "string"
          },
          "created_on": {
            "description": "新建时间",
            "type": "integer"
          },
          "deleted_on": {
            "description": "删除时间 当未被删除时,即为0,忽略此字段",
            "type": "integer"
          },
          "id": {
            "description": "ID",
            "type": "integer"
          },
          "modified_on": {
            "description": "修改时间",
            "type": "integer"
          },
          "summary": {
            "description": "摘要",
            "type": "string"
          },
          "title": {
            "description": "标题",
            "type": "string"
          },
          "user_id": {
            "description": "用户ID",
            "type": "integer"
          }
        }
      },
      "models.User": {
        "type": "object",
        "properties": {
          "avatar_src": {
            "description": "头像地址",
            "type": "string"
          },
          "company": {
            "description": "公司",
            "type": "string"
          },
          "created_on": {
            "description": "新建时间",
            "type": "integer"
          },
          "deleted_on": {
            "description": "删除时间 当未被删除时,即为0,忽略此字段",
            "type": "integer"
          },
          "description": {
            "description": "描述",
            "type": "string"
          },
          "email": {
            "description": "电子邮件",
            "type": "string"
          },
          "id": {
            "description": "ID",
            "type": "integer"
          },
          "modified_on": {
            "description": "修改时间",
            "type": "integer"
          },
          "name": {
            "description": "昵称",
            "type": "string"
          },
          "password": {
            "description": "密码",
            "type": "string"
          },
          "position": {
            "description": "职位",
            "type": "string"
          },
          "sex": {
            "description": "性别",
            "type": "integer"
          },
          "username": {
            "description": "用户名",
            "type": "string"
          },
          "web_site": {
            "description": "网站",
            "type": "string"
          }
        }
      },
      "page.Page": {
        "type": "object",
        "properties": {
          "order_by": {
            "description": "排序字段",
            "type": "string"
          },
          "page": {
            "description": "页数",
            "type": "integer"
          },
          "size": {
            "description": "条数",
            "type": "integer"
          },
          "sorter": {
            "description": "升序, 降序",
            "type": "string"
          }
        }
      },
      "v1.postIds": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "帖子IDs",
            "type": "array",
            "items": {
              "type": "integer"
            }
          }
        }
      }
    },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "Header"
    }
  }
}